Move outline file to zelm.json
This way we can differentiate Zelm projects from Elm projects and
prevent users from accidentally trying to push Zelm packages to the main
Elm package repository.

Even though we aim for bidirectional compatibility with Elm, there may
be some Zelm applications or libraries that depend on Zelm packages
which do not exist in the main Elm repository, which would mean we would
push broken packages to the Elm repository if we blindly allowed Zelm to
publish packages to the main repo.

By naming the file differently, we can hint to users that this isn't
something that should go in the normal Elm package repository.