Revert "Move outline file to zelm.json"
This reverts commit 635ed073c0b4f635e2269f9f09178247f8afe017.
If I move to a custom zelm.json file then a lot of Elm tooling will
break (the biggest casualty will be IDE integrations) which is far too
large of a price to pay.

On reflection I should also put the custom repositories bit in the
top-level ~/.elm folder anyways (because it will have global effects
rather than just a per-project effect because of the way that Elm caches
packages globally).