Fix https://github.com/Zokka-Dev/zokka-compiler/issues/8
Tail calls are being optimized incorrectly in the vanilla Elm compiler,
causing weird errors to occur when `var`s in the resulting JS are
captured in closures. This fixes that by hoisting out the while loop
body into a separate function call, which then has `var` acting
correctly within function scope.

This is inspired by how Babel deals with a similar problem (converting
`let`s into `var`s).

Note that this is really more of a PoC. I'm not sure
whether I'm generating variable names correctly and there remains the
possibility that I'm generating variables which can be overwritten by
user Elm code (i.e. I'm not 100% about the freshness of the names), but
that should be a secondary concern.

The overall approach here should be sound.