Move custom-package-server to separate repo
Cabal project kept giving me weird issues with building this project.
The API for the package server also has stabilized and it's actually
pretty useful as a moderating force to make sure that we are
conservative with changes to the package server API that we keep it in a
separate repository.

Having it in a separate repository also makes versioning and CI stuff a
little bit more manageable.